[feat] 유저가 글 작성하게 적용 * request Dto 이름 Form 으로 변경 * 로직 상 User 추가

+++ backend/src/main/java/com/NBE_4_5_2/Team5/domain/post/post/controller/ProductPostController.java
@@ -1,10 +1,12 @@
 package com.NBE_4_5_2.Team5.domain.post.post.controller;
 
-import com.NBE_4_5_2.Team5.domain.post.post.dto.request.ProductPostModifyReqBody;
+import com.NBE_4_5_2.Team5.domain.post.post.dto.request.ProductPostModifyForm;
+import com.NBE_4_5_2.Team5.domain.post.post.dto.request.ProductPostWriteForm;
 import com.NBE_4_5_2.Team5.domain.post.post.dto.response.ProductPostResponse;
-import com.NBE_4_5_2.Team5.domain.post.post.dto.request.ProductPostWriteReqBody;
 import com.NBE_4_5_2.Team5.domain.post.post.entity.ProductPost;
 import com.NBE_4_5_2.Team5.domain.post.post.service.ProductPostService;
+import com.NBE_4_5_2.Team5.domain.user.entity.User;
+import com.NBE_4_5_2.Team5.global.Rq;
 import com.NBE_4_5_2.Team5.global.dto.Empty;
 import com.NBE_4_5_2.Team5.global.dto.PageDto;
 import com.NBE_4_5_2.Team5.global.dto.RsData;
@@ -21,12 +23,14 @@ import java.io.Reader;
 @RequiredArgsConstructor
 public class ProductPostController {
     private final ProductPostService productPostService;
+    private final Rq rq;
 
     @PostMapping
-    public RsData<ProductPostResponse> createPost(@Valid @RequestBody ProductPostWriteReqBody body) {
+    public RsData<ProductPostResponse> createPost(@Valid @RequestBody ProductPostWriteForm body) {
 
-        // 작성자 체크 및 write에 넘겨주기 추가 필요
-        ProductPost post = productPostService.write(body);
+        User actor = rq.getUserIdentity();
+
+        ProductPost post = productPostService.write(actor, body);
 
         return new RsData<>(
                 "200",
@@ -65,16 +69,15 @@ public class ProductPostController {
     @PutMapping("/{id}")
     @Transactional
     public RsData<ProductPostResponse> modify(
-            @Valid @RequestBody ProductPostModifyReqBody body,
+            @Valid @RequestBody ProductPostModifyForm body,
             @PathVariable String id) {
         ProductPost post = productPostService.getPost(id);
 
-        // 수정 권한 체크 canModify
-//        if (!post.canModify()) {
-//
-//        }
+        User actor = rq.getUserIdentity();
+        if (post.canModify(actor)) {
+            productPostService.modify(post, body);
 
-        productPostService.modify(post,body);
+        }
 
         return new RsData<>(
                 "200",
@@ -86,14 +89,12 @@ public class ProductPostController {
     @DeleteMapping("/{id}")
     @Transactional
     public RsData<Empty> delete(@PathVariable String id) {
+        User actor = rq.getUserIdentity();
         ProductPost post = productPostService.getPost(id);
 
-        // 삭제 권한 체크 canDelete
-//        if (!post.canDelete()) {
-//
-//        }
-
-        productPostService.delete(post);
+        if (post.canDelete(actor)) {
+            productPostService.delete(post);
+        }
 
         return new RsData<>(
                 "200",
